lan 2

+++ app/Services/Api/Productions/Enterprise/JobService.php
@@ -6,50 +6,58 @@
  * Time: 5:04 PM
  */
 
-namespace App\Services\Api\Productions\Admin;
+namespace App\Services\Api\Productions\Enterprise;
 
 
+use App\Models\Enterprise;
 use App\Models\Job;
 use App\Services\Api\Interfaces\ManageInterface;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 
 class JobService implements ManageInterface
 {
+    private $enterprise;
+    public function __construct()
+    {
+        $this->enterprise = Enterprise::findOrFail(Auth::user()->enterprise);
+    }
 
     public function getAll()
     {
-        $jobs = Job::all();
+        $jobs = $this->enterprise->jobs;
         foreach ($jobs as $job)
         {
-            // tra ve ten doanh nghiep + avatar
-            $job->enterprise = $job->enterprise()->select('name','avatar')->first();
-            $job->skills = $job->skills()->all();
-            $job->salary = $job->salary()->first();
-            $job->positions = $job->positions()->all();
-            $job->types = $job->types()->all();
+            $job->types = $job->types()->select('name')->get();
+            $job->skills = $job->skills->select('name')->get();
+            $job->positions = $job->positions->select('name')->get();
         }
-        return $jobs;
     }
 
     public function getOne($id)
     {
-        $job = Job::findOrFail($id);
-        $job->enterprise = $job->enterprise()->select('name','avatar')->first();
-        $job->skills = $job->skills()->all();
-        $job->salary = $job->salary()->first();
-        $job->positions = $job->positions()->all();
-        $job->types = $job->types()->all();
+        $job = $this->enterprise->jobs()->where('id',$id)->first();
+        if($job == null)
+        {
+            abort_unless(false,404);
+        }
+        $job->types = $job->types()->select('name')->get();
+        $job->skills = $job->skills->select('name')->get();
+        $job->positions = $job->positions->select('name')->get();
         return $job;
     }
 
     public function getProfile($option)
     {
-        return Job::select($option)->get();
+        return $this->enterprise->jobs()->select($option)->get();
     }
 
     public function save($inputs)
     {
+
+        $inputs->enterprise_id = Auth::user()->enterprise->id;
         $job = JobService::create($inputs->all());
+
         if($inputs->hasFile('attachment'))
         {
             $job->attachment = $inputs->file('attachment')->store('public/attachment');
@@ -64,8 +72,13 @@ class JobService implements ManageInterface
     public function update($inputs, $id)
     {
         try{
+            $job = $this->enterprise->jobs()->where('id',$id)->first();
+            if($job == null)
+            {
+                abort_unless(false,404);
+            }
             $columns = Schema::getColumnListing((new Job())->getTableName());
-            $job = Job::findOrFail($id);
+
             foreach ($columns as $column)
             {
                 if($column == 'skills')
@@ -100,7 +113,12 @@ class JobService implements ManageInterface
 
     public function destroy($id)
     {
-        $job = Job::findOrFail($id);
+        $job = $this->enterprise->jobs()->where('id',$id)->first();
+        if($job == null)
+        {
+            abort_unless(false,404);
+        }
+
         $job->types()->detach();
         $job->skills()->detach();
         $job->positions()->detach();
@@ -118,6 +136,10 @@ class JobService implements ManageInterface
         foreach ($array as $item)
         {
             $job = Job::find($item);
+            if(!$job->enterprise->id != Auth::user()->enterprise->id)
+            {
+                continue;
+            }
             if($job)
             {
                 $job->types()->detach();
